requirements.txt
chromadb==0.5.18
pandas==2.2.3
protobuf==5.29.2
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.6.0
sentence_transformers==3.2.1
setuptools==65.5.0
spacy==3.8.2
google-generativeai==0.8.3
